+++ core/src/com/unciv/Constants.kt
@@ -88,6 +88,9 @@ object Constants {
 
     const val uniqueOrDelimiter = "\" OR \""
 
+    const val simulationCiv1 = "SimulationCiv1"
+    const val simulationCiv2 = "SimulationCiv2"
+
     const val dropboxMultiplayerServer = "Dropbox"
     const val uncivXyzServer = "https://uncivserver.xyz"